added rex plugin
