+++ packages/web/src/game/constants/GameConfig.ts
@@ -3,12 +3,12 @@ import Boot from "../scenes/Boot";
 import GameScene from "../scenes/Game";
 import Preload from "../scenes/Preload";
 import phaserReact from "phaser3-react";
-
+import UIPlugin from "phaser3-rex-plugins/templates/ui/ui-plugin"
 
 export const gameConfig: Phaser.Types.Core.GameConfig = {
     title: "proto",
     type: Phaser.AUTO,
-    parent: "gameElement",
+    parent: "game-parent",
     backgroundColor: 0x383838,
     dom: {
         createContainer: true,
@@ -17,7 +17,7 @@ export const gameConfig: Phaser.Types.Core.GameConfig = {
         width: "100%",
         height: "100%",
         mode: Phaser.Scale.FIT,
-        fullscreenTarget: "gameElement"
+        fullscreenTarget: "game-parent"
     },
     physics: {
         default: 'matter',
@@ -29,15 +29,21 @@ export const gameConfig: Phaser.Types.Core.GameConfig = {
     render: {
         pixelArt: true,
     },
-    // react components in phaser
     plugins: {
         global: [
+            // import directly react components into phaser
             {
                 key: 'phaser-react',
                 plugin: phaserReact,
                 start: true
+            },
+            {
+                key: 'rexUI',
+                plugin: UIPlugin,
+                mapping: 'rexUI'
             }
-        ]
+        ],
+        
     },
     scene: [Boot, Preload, GameScene]
 };
